requirements.txt
# reconx - Penetration Testing Framework
# Python 3.11+ required

# CLI framework
click>=8.1.0
typer>=0.9.0

# Async support
asyncio>=3.4.3
aiohttp>=3.8.0

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# TUI/CLI output
rich>=13.0.0
textual>=0.41.0

# HTTP requests
httpx>=0.25.0
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
